pipeline: allow to match on cookie for flow_add
OpenFlow spec (all versions) treats ADD as ADD or UPDATE,
even though MODIFY action also exists. To make things even more
bizzare, it strictly forbbids to match on cookie for ADD, but
allows the checking on MODIFY.

This is annoying, specially when simple control planes want to be
implemented (e.g. without explicit ECMP support via GROUPS).

This commit keeps the API of1x_add_flow_entry_table() as it was,
to keep backwards compatibility and be fully OF compliant, but it
opens a back-door, __of1x_add_flow_entry_table() to be able to
match in cookie _also_ during insertion.